Completed quote table ticket

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/Quote.java
@@ -7,9 +7,9 @@ public class Quote implements Entity<String> {
   private String ticker;
   private Double lastPrice;
   private Double bidPrice;
-  private Integer bidSize;
+  private Long bidSize;
   private Double askPrice;
-  private Integer askSize;
+  private Long askSize;
 
   @Override
   public String getId() {
@@ -45,11 +45,11 @@ public class Quote implements Entity<String> {
     this.bidPrice = bidPrice;
   }
 
-  public Integer getBidSize() {
+  public Long getBidSize() {
     return bidSize;
   }
 
-  public void setBidSize(Integer bidSize) {
+  public void setBidSize(Long bidSize) {
     this.bidSize = bidSize;
   }
 
@@ -61,11 +61,11 @@ public class Quote implements Entity<String> {
     this.askPrice = askPrice;
   }
 
-  public Integer getAskSize() {
+  public Long getAskSize() {
     return askSize;
   }
 
-  public void setAskSize(Integer askSize) {
+  public void setAskSize(Long askSize) {
     this.askSize = askSize;
   }
